backend/requirements.txt
absl-py
anyio
asgiref
astunparse
cachetools
certifi
cffi
charset-normalizer
click
cmake
contourpy
cryptography
cycler
DAWG-Python
Deprecated
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
docopt
exceptiongroup
fastapi
ffmpeg
filelock
flatbuffers
fonttools
fsspec
gast
google-auth
google-auth-oauthlib
google-pasta
graphviz
grpcio
h11
h5py
huggingface-hub
idna
Jinja2
joblib
jwcrypto
keras
kiwisolver
libclang
lit
llvmlite
Markdown
MarkupSafe
matplotlib
more-itertools
mpmath
mysqlclient
networkx
nltk
numba
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
oauthlib
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
opt-einsum
packaging
pandas
Pillow
pip-check
plotly
prometheus-client
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
pybind11
pycparser
pydantic
PyJWT
PyMySQL
pyparsing
python-dateutil
pytz
PyYAML
regex
requests
requests-oauthlib
rsa
safetensors
scikit-learn
scipy
six
sniffio
sqlparse
starlette
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
terminaltables
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
triton
typing_extensions
urllib3
uvicorn
Werkzeug
wrapt